+++ backend/src/indexer/block/special_transaction_outcomes.rs
@@ -9,7 +9,7 @@ use anyhow::Context;
 use concordium_rust_sdk::{
     base::contracts_common::CanonicalAccountAddress,
     id::types::AccountAddress,
-    types::{queries::BlockInfo, AbsoluteBlockHeight, ProtocolVersion, SpecialTransactionOutcome},
+    types::{queries::BlockInfo, AbsoluteBlockHeight, SpecialTransactionOutcome},
     v2,
 };
 
@@ -424,7 +424,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                             amount.micro_ccd.try_into()?,
                             block_info.block_height,
                             AccountStatementEntryType::BakerReward,
-                            block_info.protocol_version,
                             statistics,
                         )
                     })
@@ -441,7 +440,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     mint_platform_development_charge.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::FoundationReward,
-                    block_info.protocol_version,
                     statistics,
                 )?];
                 Self::Rewards(rewards)
@@ -458,7 +456,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                             amount.micro_ccd.try_into()?,
                             block_info.block_height,
                             AccountStatementEntryType::FinalizationReward,
-                            block_info.protocol_version,
                             statistics,
                         )
                     })
@@ -477,7 +474,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     foundation_charge.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::FoundationReward,
-                    block_info.protocol_version,
                     statistics,
                 )?,
                 AccountReceivedReward::prepare(
@@ -485,7 +481,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     baker_reward.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::BakerReward,
-                    block_info.protocol_version,
                     statistics,
                 )?,
             ]),
@@ -497,7 +492,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                 development_charge.micro_ccd.try_into()?,
                 block_info.block_height,
                 AccountStatementEntryType::FoundationReward,
-                block_info.protocol_version,
                 statistics,
             )?]),
             SpecialTransactionOutcome::PaydayAccountReward {
@@ -511,7 +505,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     transaction_fees.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::TransactionFeeReward,
-                    block_info.protocol_version,
                     statistics,
                 )?,
                 AccountReceivedReward::prepare(
@@ -519,7 +512,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     baker_reward.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::BakerReward,
-                    block_info.protocol_version,
                     statistics,
                 )?,
                 AccountReceivedReward::prepare(
@@ -527,7 +519,6 @@ impl PreparedSpecialTransactionOutcomeUpdate {
                     finalization_reward.micro_ccd.try_into()?,
                     block_info.block_height,
                     AccountStatementEntryType::FinalizationReward,
-                    block_info.protocol_version,
                     statistics,
                 )?,
             ]),
@@ -588,7 +579,6 @@ impl AccountReceivedReward {
         amount: i64,
         block_height: AbsoluteBlockHeight,
         transaction_type: AccountStatementEntryType,
-        protocol_version: ProtocolVersion,
         statistics: &mut Statistics,
     ) -> anyhow::Result<Self> {
         statistics.reward_stats.increment(account_address.get_canonical_address(), amount);
@@ -599,11 +589,7 @@ impl AccountReceivedReward {
                 block_height,
                 transaction_type,
             )?,
-            update_stake:           RestakeEarnings::prepare(
-                account_address,
-                amount,
-                protocol_version,
-            ),
+            update_stake:           RestakeEarnings::prepare(account_address, amount),
         })
     }
 
@@ -621,27 +607,20 @@ struct RestakeEarnings {
     canonical_account_address: CanonicalAccountAddress,
     /// Amount of CCD received as reward.
     amount:                    i64,
-    /// Protocol version belonging to the block being processed
-    protocol_version:          ProtocolVersion,
 }
 
 impl RestakeEarnings {
-    fn prepare(
-        account_address: &AccountAddress,
-        amount: i64,
-        protocol_version: ProtocolVersion,
-    ) -> Self {
+    fn prepare(account_address: &AccountAddress, amount: i64) -> Self {
         Self {
             canonical_account_address: account_address.get_canonical_address(),
             amount,
-            protocol_version,
         }
     }
 
     async fn save(&self, tx: &mut sqlx::PgTransaction<'_>) -> anyhow::Result<()> {
         // Update the account if delegated_restake_earnings is set and is true, meaning
         // the account is delegating.
-        let account_row = sqlx::query!(
+        sqlx::query!(
             "UPDATE accounts
                 SET
                     delegated_stake = CASE
@@ -655,44 +634,7 @@ impl RestakeEarnings {
         )
         .fetch_one(tx.as_mut())
         .await?;
-        if let Some(restake) = account_row.delegated_restake_earnings {
-            let bakers_expected_affected_range = if self.protocol_version > ProtocolVersion::P6 {
-                1..=1
-            } else {
-                0..=1
-            };
-            // Account is delegating.
-            if let (true, Some(pool)) = (restake, account_row.delegated_target_baker_id) {
-                // If restake is enabled and the target is a validator pool (not the passive
-                // pool) and we update the pool stake.
-                sqlx::query!(
-                    "UPDATE bakers
-                         SET pool_total_staked = pool_total_staked + $2
-                         WHERE id = $1",
-                    pool,
-                    self.amount,
-                )
-                .execute(tx.as_mut())
-                .await?
-                .ensure_affected_rows_in_range(bakers_expected_affected_range)?;
-            }
-        } else {
-            // When delegated_restake_earnings is None the account is not delegating, so it
-            // might be baking.
-            sqlx::query!(
-                "UPDATE bakers
-                    SET
-                        staked = staked + $2,
-                        pool_total_staked = pool_total_staked + $2
-                WHERE id = $1 AND restake_earnings",
-                account_row.index,
-                self.amount
-            )
-            .execute(tx.as_mut())
-            .await?
-            // An account might still earn rewards after stopping validation or delegation.
-            .ensure_affected_rows_in_range(0..=1)?;
-        }
+
         Ok(())
     }
 }