+++ frontend/src/types/generated.ts
@@ -2413,6 +2413,7 @@ export type Query = {
   importState: ImportState;
   latestChainParameters: LatestChainParameters;
   latestTransactions?: Maybe<Array<LatestTransactionResponse>>;
+  liveStablecoins: Array<StableCoin>;
   moduleReferenceEvent: ModuleReferenceEvent;
   nodeStatus?: Maybe<NodeStatus>;
   nodeStatuses: NodeStatusConnection;