+++ backend/CHANGELOG.md
@@ -4,10 +4,29 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+
+## [2.0.14] - 2025-07-10
+
+### Changed
+
+- Staked amounts fix: now updated through querying the node
+- Fix P3->P4 protocol update migration to allow for bakers pending removal to be absent from the bakers table.
+
+## [2.0.13] - 2025-07-03
+
+### Changed
+
+Database schema version: 38
+
+- Baker APY query db migration to change the function to prevent overflow FLOAT8
+
+## [2.0.12] - 2025-07-03
+
 ### Changed
 
 - New account_transaction_type enum values added `TokenUpdate`, Removed `TokenHolder` and `TokenGovernance`
 - TokenHolder and TokenGovernance events are now merged into `TokenUpdate` event.
+- The ccdscan-api now connects to the grpc node to fetch token list and details (This will be removed later).
 
 ## [2.0.11] - 2025-07-01